Chore(curriculum editor): clean up components
