--- apps/admin/src/features/Curriculum/Components/Editor/ModuleOrderColumn.tsx
@@ -1,9 +1,0 @@
-type ModuleOrderColumnProps = {};
-
-export function ModuleOrderColumn({}: ModuleOrderColumnProps) {
-  return (
-    <>
-      <p>ModuleOrderColumn</p>
-    </>
-  );
-}
