+++ apps/admin/src/features/Curriculum/Components/Editor/SortableLessonContainer.tsx
@@ -1,0 +1,88 @@
+import type { CurriculumDraft } from "@ludocode/types";
+import {
+  closestCenter,
+  DndContext,
+  MeasuringStrategy,
+  type DragEndEvent,
+} from "@dnd-kit/core";
+import {
+  SortableContext,
+  verticalListSortingStrategy,
+} from "@dnd-kit/sortable";
+import { withForm } from "../../types";
+import { EditorLesson } from "./EditorLesson";
+import { ShadowLessButton } from "../ShadowLessButton";
+import { createNewLessonTemplate } from "./templates";
+
+export const SortableLessonContainer = withForm({
+  defaultValues: {
+    modules: [] as CurriculumDraft["modules"],
+  },
+  props: {
+    moduleIndex: 0,
+  },
+  render: function Render({ form, moduleIndex }) {
+    return (
+      <form.Field name={`modules[${moduleIndex}].lessons`} mode="array">
+        {(lessonsField) => {
+          const lessons = lessonsField.state.value;
+
+          const handleDragEnd = (event: DragEndEvent) => {
+            const { active, over } = event;
+            if (!over) return;
+            if (active.id === over.id) return;
+
+            const from = lessons.findIndex((l) => l.id === active.id);
+            const to = lessons.findIndex((l) => l.id === over.id);
+
+            if (from === -1 || to === -1) return;
+
+            lessonsField.moveValue(from, to);
+          };
+
+          return (
+            <>
+              <DndContext
+                autoScroll={false}
+                measuring={{
+                  droppable: { strategy: MeasuringStrategy.Always },
+                }}
+                onDragEnd={handleDragEnd}
+                collisionDetection={closestCenter}
+              >
+                <div className="flex flex-col gap-4 p-4 w-full h-full">
+                  <SortableContext
+                    items={lessons.map((l) => l.id)}
+                    strategy={verticalListSortingStrategy}
+                  >
+                    {lessons.map((lesson, lessonIndex) => (
+                      <EditorLesson
+                        key={lesson.id}
+                        form={form}
+                        lesson={lesson}
+                        moduleIndex={moduleIndex}
+                        lessonIndex={lessonIndex}
+                      />
+                    ))}
+                  </SortableContext>
+                </div>
+              </DndContext>
+
+              <div className="w-full flex justify-end pr-4 items-center gap-4">
+                <ShadowLessButton
+                  type="button"
+                  variant="white"
+                  onClick={() =>
+                    lessonsField.pushValue(createNewLessonTemplate())
+                  }
+                >
+                  Add Lesson
+                </ShadowLessButton>
+              </div>
+            </>
+          );
+        }}
+      </form.Field>
+    );
+  },
+});
