+++ apps/admin/src/features/Curriculum/Components/Preview/ModulePreviewItem.tsx
@@ -1,25 +1,28 @@
-import type { CurriculumDraft, CurriculumDraftLesson, CurriculumDraftLessons, LessonSnap } from "@ludocode/types";
-import { CurriculumLessonPreviewItem } from "./CurriculumLessonPreviewItem";
+import type {
+  CurriculumDraftLesson,
+  CurriculumDraftLessons,
+} from "@ludocode/types";
+import { LessonPreviewItem } from "./LessonPreviewItem";
 
-type CurriculumModuleListProps = {
+type ModulePreviewItemProps = {
   title: string;
   lessons: CurriculumDraftLessons;
   onLessonClick: (lesson: CurriculumDraftLesson) => void;
   selectedLesson: CurriculumDraftLesson | null;
 };
 
-export function CurriculumModuleList({
+export function ModulePreviewItem({
   title,
   lessons,
   onLessonClick,
   selectedLesson,
-}: CurriculumModuleListProps) {
+}: ModulePreviewItemProps) {
   return (
     <div className="flex flex-col">
       <p>{title}</p>
       <div className="w-full flex flex-col gap-2 p-4">
         {lessons.map((lesson) => (
-          <CurriculumLessonPreviewItem
+          <LessonPreviewItem
             isSelected={selectedLesson?.id == lesson.id}
             onClick={() => onLessonClick(lesson)}
             title={lesson.title}
