+++ apps/admin/src/features/Curriculum/Components/Zone/CurriculumHero.tsx
