+++ apps/admin/src/hooks/Queries/Mutations/useUpdateCourse.tsx
@@ -1,6 +1,5 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { qk } from "@/hooks/Queries/Definitions/qk.ts";
-import type { LudoCourse } from "@ludocode/types/Catalog/LudoCourse.ts";
 import { mutations } from "@/hooks/Queries/Definitions/mutations";
 import type { CurriculumDraft } from "@ludocode/types";
