+++ apps/admin/src/routes/_app/curriculum/route.tsx
@@ -1,5 +1,4 @@
 import { CurriculumLayout } from "@/layouts/Curriculum/CurriculumLayout";
-import { AdminHubLayout } from "@/layouts/Hub/AdminHubLayout";
 import { createFileRoute } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/_app/curriculum")({
